mypy: CI tip denetimi kapısını onar (85 hata -> 0)
PR CI'ında 'Lint & Type Check' başarısızdı. Sebep: sandbox'ta mypy
kurulu olmadığı için bu kapıyı hiç koşamamıştım; ruff temizdi ama mypy
85 hata veriyordu. Baz commit'e karşı doğruladım: main TEMİZ ('no issues
found'), yani 85 hatanın tamamı bu daldan geldi.

Düzeltmelerin çoğu tip annotation eksiği:
- self_learning_scaling (22), signature (28), memory_hierarchy (8):
  dict literal'den dar tip çıkarımı; Dict[str, Any] ve açık annotation
- signature._torch(): dönüş tipi eklendi -> 19 union-attr hatası düştü
- _distractors protected_nodes: Sequence -> Collection (sadece üyelik
  testi yapılıyor, set geçiliyordu)
- _percentiles / read_latency_by_tier / ram: ilan edilen tipler gerçeği
  yansıtmıyordu ('n' int, diğerleri float)
- priority_optimization: ExplorationEngine(**weights) -> açık kwargs
- twt_results, seed_statistics, engineering_audit, semantic_extraction,
  depth_diagnosis, multi_environment: annotation + bool()/float() daraltma

Tip denetimi iki GERÇEK bug yakaladı (yalnız stil değil):
1. semantik_ayiklayici: tip-tabanlı özellik döngüsünde Ozellik(value=...)
   yanlış değişkeni (onceki dongunun 'deger'i) kullaniyordu; artik
   tip_degeri.
2. semantik_ayiklayici: ozellik dongusunde ozellik_id kontrol edilip
   ONCEKI dongunun artigi entity_id ile yaziliyordu -> ozellikler yanlis
   varliga yazilabilirdi. Simdi ozellik_id kullaniliyor.

mypy: 130 dosyada 'no issues found'. ruff check . temiz.
Test: 896 passed, 0 failed (davranış değişmedi).

+++ hga/evaluation/priority_ablation.py
@@ -114,7 +114,7 @@ def kendall_tau(a: Sequence[float], b: Sequence[float]) -> float:
     payda = ((toplam - bag_a) * (toplam - bag_b)) ** 0.5
     if payda == 0:
         return 1.0
-    return round((uyumlu - uyumsuz) / payda, 8)
+    return float(round((uyumlu - uyumsuz) / payda, 8))
 
 
 def spearman_rho(a: Sequence[float], b: Sequence[float]) -> float:
@@ -131,7 +131,7 @@ def spearman_rho(a: Sequence[float], b: Sequence[float]) -> float:
     nb = sum((y - mb) ** 2 for y in rb) ** 0.5
     if na == 0 or nb == 0:
         return 1.0
-    return round(pay / (na * nb), 8)
+    return float(round(pay / (na * nb), 8))
 
 
 def _oran(pay: int, payda: int) -> float:
